REFACTOR: move test
